+++ sdk/dapp-kit/src/hooks/wallet/useConnectionStatus.ts
@@ -1,11 +1,11 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useDAppKitStore } from '../useDAppKitStore.js';
+import { useWalletStore } from './useWalletStore.js';
 
 /**
  * Retrieves the
  */
 export function useConnectionStatus() {
-	return useDAppKitStore((state) => state.connectionStatus);
+	return useWalletStore((state) => state.connectionStatus);
 }
