+++ sdk/dapp-kit/src/hooks/wallet/useCurrentAccount.ts
@@ -1,11 +1,11 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useDAppKitStore } from '../useDAppKitStore.js';
+import { useWalletStore } from './useWalletStore.js';
 
 /**
  * Retrieves the wallet account that is currently selected, if one exists.
  */
 export function useCurrentAccount() {
-	return useDAppKitStore((state) => state.currentAccount);
+	return useWalletStore((state) => state.currentAccount);
 }
