+++ sdk/dapp-kit/src/walletStore.ts
@@ -1,0 +1,102 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { Wallet, WalletAccount, WalletWithRequiredFeatures } from '@mysten/wallet-standard';
+import { createStore } from 'zustand';
+import type { StateStorage } from 'zustand/middleware';
+import { persist, createJSONStorage } from 'zustand/middleware';
+
+export type WalletActions = {
+	setAccountSwitched: (selectedAccount: WalletAccount) => void;
+	setWalletConnected: (
+		wallet: WalletWithRequiredFeatures,
+		selectedAccount: WalletAccount | null,
+	) => void;
+	setWalletDisconnected: () => void;
+	setWalletRegistered: (updatedWallets: WalletWithRequiredFeatures[]) => void;
+	setWalletUnregistered: (
+		updatedWallets: WalletWithRequiredFeatures[],
+		unregisteredWallet: Wallet,
+	) => void;
+};
+
+export type WalletStore = ReturnType<typeof createWalletStore>;
+
+export type StoreState = {
+	wallets: WalletWithRequiredFeatures[];
+	currentWallet: WalletWithRequiredFeatures | null;
+	currentAccount: WalletAccount | null;
+	lastConnectedAccountAddress: string | null;
+	lastConnectedWalletName: string | null;
+	connectionStatus: 'disconnected' | 'connecting' | 'connected';
+} & WalletActions;
+
+export type WalletConfiguration = {
+	wallets: WalletWithRequiredFeatures[];
+	storage: StateStorage;
+	storageKey: string;
+};
+
+export function createWalletStore({ wallets, storage, storageKey }: WalletConfiguration) {
+	return createStore<StoreState>()(
+		persist(
+			(set, get) => ({
+				wallets,
+				currentWallet: null,
+				currentAccount: null,
+				lastConnectedAccountAddress: null,
+				lastConnectedWalletName: null,
+				connectionStatus: 'disconnected',
+				setWalletConnected: (wallet, selectedAccount) => {
+					set(() => ({
+						currentWallet: wallet,
+						currentAccount: selectedAccount,
+						lastConnectedWalletName: wallet.name,
+						lastConnectedAccountAddress: selectedAccount?.address,
+						connectionStatus: 'connected',
+					}));
+				},
+				setWalletDisconnected: () => {
+					set(() => ({
+						currentWallet: null,
+						currentAccount: null,
+						lastConnectedWalletName: null,
+						lastConnectedAccountAddress: null,
+						connectionStatus: 'disconnected',
+					}));
+				},
+				setAccountSwitched: (selectedAccount) => {
+					set(() => ({
+						currentAccount: selectedAccount,
+						lastConnectedAccountAddress: selectedAccount.address,
+					}));
+				},
+				setWalletRegistered: (updatedWallets) => {
+					set(() => ({ wallets: updatedWallets }));
+				},
+				setWalletUnregistered: (updatedWallets, unregisteredWallet) => {
+					if (unregisteredWallet === get().currentWallet) {
+						set(() => ({
+							wallets: updatedWallets,
+							currentWallet: null,
+							currentAccount: null,
+							lastConnectedWalletName: null,
+							lastConnectedAccountAddress: null,
+							connectionStatus: 'disconnected',
+						}));
+					} else {
+						set(() => ({ wallets: updatedWallets }));
+					}
+				},
+			}),
+			{
+				name: storageKey,
+				storage: createJSONStorage(() => storage),
+				partialize: ({ lastConnectedWalletName, lastConnectedAccountAddress }) => ({
+					lastConnectedWalletName,
+					lastConnectedAccountAddress,
+				}),
+			},
+		),
+	);
+}
