+++ sdk/dapp-kit/test/hooks/useConnectWallet.test.tsx
@@ -8,7 +8,7 @@ import {
 	useCurrentWallet,
 	useCurrentAccount,
 } from 'dapp-kit/src';
-import { createDAppKitProviderContextWrapper, registerMockWallet } from '../test-utils.js';
+import { createWalletProviderContextWrappe, registerMockWallet } from '../test-utils.js';
 import { WalletAlreadyConnectedError } from 'dapp-kit/src/errors/walletErrors.js';
 import type { Mock } from 'vitest';
 
@@ -16,7 +16,7 @@ describe('useConnectWallet', () => {
 	test('throws an error when connecting to a wallet when a connection is already active', async () => {
 		const { unregister, mockWallet } = registerMockWallet({ walletName: 'Mock Wallet 1' });
 
-		const wrapper = createDAppKitProviderContextWrapper();
+		const wrapper = createWalletProviderContextWrappe();
 		const { result } = renderHook(() => useConnectWallet(), { wrapper });
 
 		result.current.mutate({ wallet: mockWallet });
@@ -33,7 +33,7 @@ describe('useConnectWallet', () => {
 	test('throws an error when a user fails to connect their wallet', async () => {
 		const { unregister, mockWallet } = registerMockWallet({ walletName: 'Mock Wallet 1' });
 
-		const wrapper = createDAppKitProviderContextWrapper();
+		const wrapper = createWalletProviderContextWrappe();
 		const { result } = renderHook(
 			() => ({
 				connectWallet: useConnectWallet(),
@@ -62,7 +62,7 @@ describe('useConnectWallet', () => {
 	test('connecting to a wallet works successfully', async () => {
 		const { unregister, mockWallet } = registerMockWallet({ walletName: 'Mock Wallet 1' });
 
-		const wrapper = createDAppKitProviderContextWrapper();
+		const wrapper = createWalletProviderContextWrappe();
 		const { result } = renderHook(
 			() => ({
 				connectWallet: useConnectWallet(),
