+++ sdk/dapp-kit/test/hooks/useDisconnectWallet.test.tsx
@@ -9,12 +9,12 @@ import {
 	useCurrentWallet,
 	useConnectionStatus,
 } from 'dapp-kit/src';
-import { createDAppKitProviderContextWrapper, registerMockWallet } from '../test-utils.js';
+import { createWalletProviderContextWrappe, registerMockWallet } from '../test-utils.js';
 import { WalletNotConnectedError } from 'dapp-kit/src/errors/walletErrors.js';
 
 describe('useDisconnectWallet', () => {
 	test('that an error is thrown when trying to disconnect with no active connection', async () => {
-		const wrapper = createDAppKitProviderContextWrapper();
+		const wrapper = createWalletProviderContextWrappe();
 		const { result } = renderHook(() => useDisconnectWallet(), { wrapper });
 
 		result.current.mutate();
@@ -25,7 +25,7 @@ describe('useDisconnectWallet', () => {
 	test('that disconnecting works successfully', async () => {
 		const { unregister, mockWallet } = registerMockWallet({ walletName: 'Mock Wallet 1' });
 
-		const wrapper = createDAppKitProviderContextWrapper();
+		const wrapper = createWalletProviderContextWrappe();
 		const { result } = renderHook(
 			() => ({
 				connectWallet: useConnectWallet(),
